+++ example/showcase/src/app/functions.rs
