--- example/showcase/src/job/mod.rs
@@ -1,3 +1,0 @@
-pub mod functions;
-pub mod join;
-pub mod simple;
