+++ example/showcase/src/main.rs
@@ -3,9 +3,9 @@ extern crate log;
 #[macro_use]
 extern crate rlink_derive;
 
+mod app;
 mod buffer_gen;
-mod job;
 
 pub fn main() {
-    rlink::api::env::execute("showcase", crate::job::join::JoinStreamApp {});
+    rlink::api::env::execute("showcase", crate::app::JoinStreamApp {});
 }
